requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
openai==1.17.0
python-telegram-bot==13.15
python-dotenv==1.0.1
tiktoken==0.6.0
aiohttp==3.9.3
transformers==4.39.3
torch>=2.1.0
scikit-learn>=1.4.0
requests==2.31.0